Refactor `worms_packet`, `packet_code`, and associated enums to use `uint16_t`, improve consistency by collapsing single-line methods across multiple classes, standardize method formatting, and enhance encoding logic with utility functions and safer operations.

+++ include/packet_flags.hpp
@@ -7,23 +7,26 @@
 
 #include <cstdint>
 
-namespace worms_server {
-    enum class PacketFlags : uint32_t {
-        Value0       = 1 << 0,
-        Value1       = 1 << 1,
-        Value2       = 1 << 2,
-        Value3       = 1 << 3,
-        Value4       = 1 << 4,
-        Value10      = 1 << 10,
+namespace worms_server
+{
+    enum class PacketFlags : uint16_t
+    {
+        Value0      = 1 << 0,
+        Value1      = 1 << 1,
+        Value2      = 1 << 2,
+        Value3      = 1 << 3,
+        Value4      = 1 << 4,
+        Value10     = 1 << 10,
         DataLength  = 1 << 5,
-        Data         = 1 << 6,
-        Error        = 1 << 7,
-        Name         = 1 << 8,
+        Data        = 1 << 6,
+        Error       = 1 << 7,
+        Name        = 1 << 8,
         SessionInfo = 1 << 9,
     };
 
     // Helper function to test flags
-    static constexpr bool HasFlag(const uint32_t value, const PacketFlags flag) {
+    static constexpr bool HasFlag(const uint32_t value, const PacketFlags flag)
+    {
         return (value & static_cast<uint32_t>(flag)) != 0;
     }
 } // namespace worms_server
